+++ lua/plugins/config/lualine_config/slanted_gaps.lua
@@ -50,7 +50,7 @@ end
 
 return {
 	options = {
-		theme = 'catppuccin',
+		theme = 'auto',
 		component_separators = '',
 		section_separators = { left = '', right = '' },
 	},